backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.13.3
python-multipart==0.0.9
python-dotenv==1.0.1
paho-mqtt==2.1.0
asyncua==1.1.0
httpx==0.27.2
redis==5.1.0
orjson==3.10.7
fpdf2==2.8.1
numpy==2.1.2
tenacity==9.0.0
loguru==0.7.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.9.0
email-validator==2.2.0
apscheduler==3.10.4
pymssql==2.3.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
aiosmtplib==2.0.2
psutil==5.9.8
